WEEK 7/requirements.txt
langchain
faiss-cpu
sentence-transformers
pypdf
python-docx
pandas
tqdm